Only filter transition replacement sources
Prior to this commit, we were filtering both source and destination in a
transition replacement operation.  This allowed some bad transitions to
remain after the operation completed.  For example, the
Needles-->DowntownWest journey ("NtD") contains a lot of relative
transitions that we filter out.  If we replace another journey XtY with
NtD, we correctly select transitions from NtD's filtered list so that we
don't copy from a relative transition.

When we add the reverse route (replace DtN with YtX), we want to replace
*every* transition from DowntownWest-->Needles, including the relative
ones and ones filtered out for any other reason.  When the player takes
a DowntownWest to Needles exit, he should always be taken to X, never to
Needles.

After this commit, we only use the filtered list when selecting
transitions to copy *from*.  The unfiltered list is used to determine
which transitions to write *to*.